pyproject.toml
[tool.poetry]
name = "packratt"
version = "0.1.0"
description = ""
authors = ["Simon Perkins <simon.perkins@gmail.com>"]
include = ["conf/registry.yaml", "conf/registry-schema.yaml"]

[tool.poetry.dependencies]
python = "^3.6"
requests = "^2.23.0"
pytest = {version = "^5.4.2", optional = true, extras = ["testing"]}
pytest-flake8 = {version = "^1.0.6", optional = true, extras = ["testing"]}
appdirs = "^1.4.4"
jsonschema = "^3.2.0"
pyyaml = "^5.3.1"


[tool.poetry.extras]
testing = ["pytest", "pytest-flake8"]

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
packratt = 'packratt.application:run'
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

